requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
prophet
google-generativeai
fpdf
tabulate